+++ packages/heroku-pipelines/test/commands/pipelines/create.js
@@ -5,25 +5,56 @@ let nock = require('nock')
 let cmd = require('../../../commands/pipelines/create')
 
 describe('pipelines:create', function () {
+  let heroku, coupling, pipeline
+
   beforeEach(function () {
     cli.mockConsole()
+    heroku = nock('https://api.heroku.com')
+    coupling = { id: '0123', stage: 'production' }
+
+    heroku.post('/pipeline-couplings')
+    .reply(201, coupling)
   })
 
-  it('displays the pipeline name and app stage', function () {
-    let pipeline = {name: 'example', id: '0123'}
-    let coupling = { id: '0123', stage: 'production' }
+  context('when not specifying ownership', function () {
+    beforeEach(function () {
+      pipeline = {name: 'example', id: '0123', owner: { id: '1234-567', type: 'user' }}
 
-    let heroku = nock('https://api.heroku.com')
+      heroku
+      .get('/account')
+      .reply(200, { id: '1234-567' })
       .post('/pipelines')
       .reply(201, pipeline)
-      .post('/pipeline-couplings')
-      .reply(201, coupling)
+    })
 
-    return cmd.run({app: 'example', args: {name: 'example'}, flags: {stage: 'production'}})
-    .then(function () {
-      cli.stderr.should.contain('Creating example pipeline... done')
-      cli.stderr.should.contain('Adding example to example pipeline as production... done')
-      heroku.done()
+    it('displays the pipeline name and app stage', function () {
+      return cmd.run({app: 'example', args: {name: 'example'}, flags: {stage: 'production'}})
+      .then(function () {
+        cli.stderr.should.contain('Creating example pipeline... done')
+        cli.stderr.should.contain('Adding example to example pipeline as production... done')
+        heroku.done()
+      })
+    })
+  })
+
+  context('when specifying a team as owner', function () {
+    beforeEach(function () {
+      pipeline = {name: 'example', id: '0123', owner: { id: '89-0123-456', type: 'team' }}
+
+      heroku
+      .get('/teams/my-team')
+      .reply(200, { id: '89-0123-456' })
+      .post('/pipelines')
+      .reply(201, pipeline)
+    })
+
+    it('displays the pipeline name and app stage', function () {
+      return cmd.run({app: 'example', args: {name: 'example'}, flags: {stage: 'production', team: 'my-team'}})
+      .then(function () {
+        cli.stderr.should.contain('Creating example pipeline... done')
+        cli.stderr.should.contain('Adding example to example pipeline as production... done')
+        heroku.done()
+      })
     })
   })
 })